requirements_livekit.txt
# LiveKit Alice Agent
livekit-agents>=1.0.0
livekit-plugins-openai
livekit-plugins-silero  
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-turn-detector

# Våra verktyg och AI
requests>=2.31.0
duckduckgo-search>=6.1.0
ollama>=0.3.0
python-dotenv>=1.0.0